+++ src/KeyConnector/Services/CertificateProviders/AzureStorageCertificateProviderService.cs
@@ -3,7 +3,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.CertificateProviders
 {
     public class AzureStorageCertificateProviderService : ICertificateProviderService
     {
