+++ src/KeyConnector/Services/CertificateProviders/FilesystemCertificateProviderService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.CertificateProviders
 {
     public class FilesystemCertificateProviderService : ICertificateProviderService
     {
