+++ src/KeyConnector/Services/CertificateProviders/ICertificateProviderService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.CertificateProviders
 {
     public interface ICertificateProviderService
     {
