+++ src/KeyConnector/Services/CertificateProviders/StoreCertificateProviderService.cs
@@ -2,7 +2,7 @@
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.CertificateProviders
 {
     public class StoreCertificateProviderService : ICertificateProviderService
     {
