+++ src/KeyConnector/Services/Crypto/CryptoFunctionService.cs
@@ -2,7 +2,7 @@
 using System.Security.Cryptography;
 using System.Threading.Tasks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.Crypto
 {
     public class CryptoFunctionService : ICryptoFunctionService
     {
