+++ src/KeyConnector/Services/Crypto/ICryptoFunctionService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Security.Cryptography;
 using System.Threading.Tasks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.Crypto
 {
     public interface ICryptoFunctionService
     {
