+++ src/KeyConnector/Services/RsaKey/RsaHealthCheckService.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 
-namespace Bit.KeyConnector.Services
+namespace Bit.KeyConnector.Services.RsaKey
 {
     public class RsaHealthCheckService : IHealthCheck
     {
